requirement.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
comtypes==1.1.10
cryptography==3.4.7
decorator==4.4.2
decorators==2.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.44.0
django-cors-headers==3.8.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
email-validator==1.3.1
idna==2.10
Pillow==8.2.0
pipenv==2021.5.29
PyJWT==2.1.
python-decouple==3.4
pyparsing==3.0.9
pypiwin32==223
python3-openid==3.2.0
pytz==2021.1
pywin32==301
redis==4.3.5
requests==2.25.1
requests-oauthlib==1.3.0
shell==1.0.1
sqlparse==0.4.2
typing-extensions==3.10.0.0
urllib3==1.26.5
validate-email==1.3
virtualenv==20.4.7
whitenoise==5.1.0